Preproces: move reusable functions to separate file
